Update script.js
I've removed the inline onclick events and added them here. There was an issue loading the images to select from on starting the game as well as an issue loading the style and javascript sheets, at least using Brave browser. I've edited the innerHTML so the elements are being created instead. I've kept in all the console logs, but they can be removed also. Hope this is of some help