Update index.html
I've removed the inline onclick functions and added them to the JS instead. There was an issue loading the script and stylesheet, at least on Brave browser, so I fixed those problems. There was also an issue loading the images to select from at the start of the game, they wouldn't show, that's been fixed also. I hope this is of some help